services/timelog/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "timelog-service"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "consulting-pm-shared",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.0.0",
    "redis[hiredis]>=5.0.0",
    "structlog>=24.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
